refactored to avoid cyclic dependencys, fixed token expiration.
